+++ CHANGELOG.md
@@ -3,14 +3,21 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 # [v1.0.2] (UnReleased)
 
+### Notes
 This version supports OneView appliances with version 4.2 using the OneView REST API version 1000.
 
+- Support for Update Group in Logical Enclosure.
 - Support for Update Compliance in Logical Interconnect.
 - Usecase for synergy infrastructure provisioning with Network.
 
 ### Oneview Features supported
-- Network Set
-- Storage System
+- Enclosure Group
+- Ethernet network
+- FC network
+- Interconnect Type
+- Logical enclosure
+- Network set
+- Storage system
 - Storage volume attachment
 
 ### Image Streamer Features supported